requirements.txt
django==3.2.5
django-extensions==3.1.3
djangorestframework==3.12.4